Externalize the versioning logic and custom Gradle scripts and also change the slf4j dependency
