+++ build.gradle.kts
@@ -1,5 +1,4 @@
 import de.thetaphi.forbiddenapis.gradle.CheckForbiddenApis
-import java.util.Properties
 
 plugins {
 	java
@@ -34,63 +33,22 @@ dependencyManagement {
     }
 }
 val checkstyleVersion = "10.17.0"
-val dependencyArchunitVersion = "1.3.0"
+val dependencySlf4jVersion = "2.0.13"
 
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
+    implementation("org.slf4j:slf4j-api:$dependencySlf4jVersion")
 
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
 
-    testImplementation("com.tngtech.archunit:archunit:$dependencyArchunitVersion")
+
     testImplementation("org.springframework.boot:spring-boot-starter-test")
     testImplementation("org.junit.jupiter:junit-jupiter")
 }
 
-class ApplicationVersion(
-    private val major: Int,
-    private val minor: Int,
-    private val patch: Int,
-    private val release: Boolean
-) {
-    private fun getRelease(): String {
-        return if (release) "" else "-SNAPSHOT"
-    }
-
-    fun getVersion(): String {
-        return "$major.$minor.$patch" + getRelease()
-    }
-}
-
-val loadVersion: () -> String = {
-    val versionPropertiesFile = File("src/main/resources/version.properties")
-
-    if (!versionPropertiesFile.exists()) {
-        throw Exception("No version.properties file found")
-    }
-
-    val versionProperties = Properties()
-
-    versionPropertiesFile.inputStream().use { stream ->
-        versionProperties.load(stream)
-    }
-
-    val major = versionProperties.getProperty("major")?.toInt() ?: 0
-    val minor = versionProperties.getProperty("minor")?.toInt() ?: 0
-    val patch = versionProperties.getProperty("patch")?.toInt() ?: 0
-    val release = versionProperties.getProperty("release")?.toBoolean() ?: false
-
-    val version = ApplicationVersion(major, minor, patch, release)
-    version.getVersion()
-}
-
-version = loadVersion()
-
-tasks.withType<Jar> {
-    manifest {
-        attributes["Implementation-Version"] = version
-    }
-}
+// Loads the versioning logic, and the custom Gradle tasks
+apply(from = "versioning.gradle.kts")
 
 tasks.withType<Test> {
     useJUnitPlatform()
@@ -113,87 +71,6 @@ tasks.named<DefaultTask>("checkstyleTest").configure {
 
 tasks.named("check").configure {
 	dependsOn(tasks.named("forbiddenApisMain"))
-}
-
-tasks.register("majorVersionUpdate") {
-    group = "versioning"
-    description = "Bump to next major version"
-
-    doFirst {
-        val versionFile = File("src/main/resources/version.properties")
-        val properties = Properties()
-
-        versionFile.inputStream().use { stream ->
-            properties.load(stream)
-        }
-
-        properties.setProperty("major", (properties.getProperty("major").toInt() + 1).toString())
-        properties.setProperty("minor", "0")
-        properties.setProperty("patch", "0")
-
-        versionFile.outputStream().use { properties.store(it, null) }
-    }
-}
-
-tasks.register("minorVersionUpdate") {
-    group = "versioning"
-    description = "Bump to next minor version"
-
-    doFirst {
-        val versionFile = file("src/main/resources/version.properties")
-        val properties = Properties()
-        versionFile.inputStream().use { properties.load(it) }
-
-        properties.setProperty("minor", (properties.getProperty("minor").toInt() + 1).toString())
-        properties.setProperty("patch", "0")
-
-        versionFile.outputStream().use { properties.store(it, null) }
-    }
-}
-
-tasks.register("patchVersionUpdate") {
-    group = "versioning"
-    description = "Bump to next patch version"
-
-    doFirst {
-        val versionFile = file("src/main/resources/version.properties")
-        val properties = Properties()
-        versionFile.inputStream().use { properties.load(it) }
-
-        properties.setProperty("patch", (properties.getProperty("patch").toInt() + 1).toString())
-
-        versionFile.outputStream().use { properties.store(it, null) }
-    }
-}
-
-tasks.register("releaseVersion") {
-    group = "versioning"
-    description = "Make the version a release"
-
-    doFirst {
-        val versionFile = file("src/main/resources/version.properties")
-        val properties = Properties()
-        versionFile.inputStream().use { properties.load(it) }
-
-        properties.setProperty("release", "true")
-
-        versionFile.outputStream().use { properties.store(it, null) }
-    }
-}
-
-tasks.register("preReleaseVersion") {
-    group = "versioning"
-    description = "Make the version a pre release"
-
-    doFirst {
-        val versionFile = file("src/main/resources/version.properties")
-        val properties = Properties()
-        versionFile.inputStream().use { properties.load(it) }
-
-        properties.setProperty("release", "false")
-
-        versionFile.outputStream().use { properties.store(it, null) }
-    }
 }
 
 idea {
